[AIRFLOW-5255] Move GCP SQL to core (#5861)
This commit moves GCP SQL from contrib to core.
For more information check AIP-21.
